+++ front-end/src/components/HomeNav.vue
@@ -9,7 +9,13 @@
         <router-link to="/search">Search</router-link>
       </li>
       <li>
+        <router-link to="/apartments">Sponsored Apartments</router-link>
+      </li>
+      <li>
         <a :href="loginUrl">Login</a>
+      </li>
+      <li>
+        <a :href="singinUrl">Singin</a>
       </li>
     </ul>
   </div>
@@ -20,6 +26,7 @@ export default {
   data() {
     return {
       loginUrl: "http://localhost:8000/login", // Memorizzazione dell'URL di login in una variabile
+      singinUrl: "http://localhost:8000/register", // Memorizzazione dell'URL di singin in una variabile
     };
   },
 };
