+++ front-end/src/main.js
@@ -1,26 +1,10 @@
 import { createApp } from 'vue'
 import './style.css'
-// import App from './App.vue'
-import { createRouter , createWebHashHistory} from 'vue-router'
-
-
 import MainLayout  from './layouts/MainLayout.vue'
-import HomePage from './pages/HomePage.vue'
-import AdvancedSearch from './pages/AdvancedSearch.vue'
+// import App from './App.vue'
 
 
-
-
-const routes = [
-    {path: '/', component: HomePage} ,
-    {path: '/search', component: AdvancedSearch}  ,
- ]
-
- const router = createRouter({
-    history: createWebHashHistory(),
-    routes,
-})
-
+import { router } from './router';
 
 
 const app = createApp(MainLayout)
