+++ front-end/src/pages/Apartments.vue
@@ -1,0 +1,5 @@
+<template>
+  <h1>Apartments</h1>
+</template>
+
+<script></script>
